+++ lib/screens/menu/your menu/edit menu item.dart
@@ -56,7 +56,7 @@ class _EditMenuItemState extends State<EditMenuItem> {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             SizedBox(
-              width: MediaQuery.of(context).size.width * 0.5,
+              width: MediaQuery.of(context).size.width * 0.4,
               child: TextFormField(
                 controller: toppingNamecontroller,
                 style: const TextStyle(color: Colors.white),
@@ -111,6 +111,8 @@ class _EditMenuItemState extends State<EditMenuItem> {
                 },
               ),
             ),
+            IconButton(onPressed: removeToppingWidget, icon: Icon(Icons.delete,color: orangeColor(),),),
+
           ],
         ),
         const SizedBox(height:10),
@@ -122,6 +124,12 @@ class _EditMenuItemState extends State<EditMenuItem> {
   void addToppingWidget(){
     setState(() {
       Toppings.add(topping());
+    });
+  }
+
+  void removeToppingWidget(){
+    setState(() {
+      Toppings.removeAt(0);
     });
   }
 
@@ -297,12 +305,52 @@ class _EditMenuItemState extends State<EditMenuItem> {
               color: Colors.white,
             ),
             onPressed: () async {
-              await DeleteMenuItem();
-              ScaffoldSnackbar.of(context).show('Menu item deleted successfully');
-              Navigator.pop(context,'true');
-              setState(() {
-                _isLoading = false;
-              });
+              showDialog(
+                  context: context,
+                  builder:(BuildContext context) {
+                    return AlertDialog(
+                      backgroundColor: cardColor(),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(10.0),
+                      ),
+                      title:Text('key_Delete_Item'.tr, style:cardTextStyle18()),
+                      content: Text('key_are_you_sure_delete'.tr, style:mTextStyle14() ),
+                      actions: [
+                        TextButton(
+                          style: TextButton.styleFrom(
+                            backgroundColor: fontColor(),
+                            shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(8.0),
+                            ),
+                          ),
+                          child: Text('key_CANCEL'.tr, style:  mTextStyle14()),
+                          onPressed: () {
+                            Navigator.of(context).pop(); // Close the dialog
+                          },
+                        ),
+                        TextButton(
+                          style: TextButton.styleFrom(
+                            backgroundColor: orangeColor(),
+                            shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(8.0),
+                            ),
+                          ),
+                          child: Text('key_Proceed'.tr, style: mTextStyle14(),),
+                          onPressed: () async {
+                            Navigator.of(context).pop(); // Close the dialog
+                            await DeleteMenuItem();
+                            ScaffoldSnackbar.of(context).show('Menu item deleted successfully');
+                            Navigator.pop(context,'true');
+                            setState(() {
+                              _isLoading = false;
+                            });
+                          },
+                        ),
+                      ],
+                    );
+                  }
+              );
+
             },
           )
         ],
@@ -469,7 +517,7 @@ class _EditMenuItemState extends State<EditMenuItem> {
                                 Padding(
                                   padding: const EdgeInsets.all(8.0),
                                   child: Text(
-                                    'key_Veg_Only'.tr,
+                                    _switchValue ? 'key_Veg'.tr :'key_Non_Veg'.tr,
                                     textAlign: TextAlign.center,
                                     style: inputTextStyle16(),
                                   ),
@@ -479,6 +527,7 @@ class _EditMenuItemState extends State<EditMenuItem> {
                                   child: CupertinoSwitch(
                                       thumbColor: Colors.black,
                                       activeColor: Colors.green,
+                                      trackColor:Colors.red,
                                       value: _switchValue,
                                       onChanged: (bool? value) {
                                         setState(() {
